// src/Juniper.TypeScript/@juniper-lib/threejs/widgets/ScreenModeToggleButton.d.ts
import { ScreenMode } from "../ScreenMode";
import { ButtonFactory } from "./ButtonFactory";
import { ToggleButton } from "./ToggleButton";
export declare class ScreenModeToggleButton extends ToggleButton {
    readonly mode: ScreenMode;
    constructor(buttons: ButtonFactory, mode: ScreenMode);
}
//# sourceMappingURL=ScreenModeToggleButton.d.ts.map